scripts/buildtools/pyproject.toml
[tool.poetry]
name = "buildtools"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.scripts]
buildtools = "buildtools.main:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
typer = "^0.9.0"
loguru = "^0.7.0"
orjson = "^3.9.2"
toml = "^0.10.2"
pygithub = "^1.59.0"
pydantic = "1.10"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
pytest-mock = "^3.10.0"
ruff = "^0.0.278"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
